refactor(ModmailConversationAuthor): convert to TS
