+++ src/objects/RedditContent.d.ts
@@ -5,6 +5,9 @@ export default class RedditContent<T> extends Promise<T> {
   created: number;
   id: string;
   name: string;
+  protected _r: Snoowrap;
+  protected _fetch?: boolean;
+  protected _hasFetched: boolean;
 
   constructor(options: any, _r: Snoowrap, _hasFetched: boolean);
   fetch(): Promise<T>;
